Make AFS deployment more secure

+++ deploy.sh
@@ -1,9 +1,5 @@
 #!/usr/bin/env bash
 
-# This script copies the GitHub Pages site stanford-dancebreak.github.io to dancebreak.stanford.edu.
-# It should be placed at /afs/ir/group/dancebreak/cgi-bin
-# It can be triggered by a POST request to dancebreak.stanford.edu/cgi-bin/deploy.sh.
-
 echo "Content-Type: text/plain"
 echo ""
 
@@ -14,9 +10,17 @@ then
 fi
 
 cd ..
+TOKEN=$(cat access/access-token.txt)
+if [[ "$HTTP_X_AUTHORIZE" != "$TOKEN" ]]
+then
+  echo "Unauthorized."
+  exit 1
+fi
+TAR=$(mktemp)
 DIR=$(mktemp -d)
-echo $DIR
-wget --mirror --page-requisites https://stanford-dancebreak.github.io -P $DIR
-ls $DIR/stanford-dancebreak.github.io
-flock -x WWW rsync -avh $DIR/stanford-dancebreak.github.io/ WWW/ --delete
+cat > $TAR
+tar -xvf $TAR -C $DIR
+ls $DIR
+rm $TAR
+flock -x WWW rsync -avh $DIR/ WWW/ --delete
 rm -rf $DIR
